refactored test_seaborn_table
